aat-dogbreed-app/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pillow==10.4.0
tensorflow==2.16.1
python-multipart==0.0.9
pydantic==2.9.2